Moved common AuthClient tests to test.authclient.ts
