+++ test/test.compute.ts
@@ -19,7 +19,6 @@ import * as nock from 'nock';
 import * as sinon from 'sinon';
 import {Compute, gcpMetadata} from '../src';
 import {
-  NoOpEncodedLocations,
   SERVICE_ACCOUNT_LOOKUP_ENDPOINT,
   TrustBoundaryData,
 } from '../src/auth/trustboundary';
@@ -272,14 +271,9 @@ describe('compute', () => {
 
     const MOCK_ACCESS_TOKEN = 'abc123';
     const MOCK_AUTH_HEADER = `Bearer ${MOCK_ACCESS_TOKEN}`;
-    const SERVICE_ACCOUNT_EMAIL = 'service-account@example.com';
     const EXPECTED_TB_DATA: TrustBoundaryData = {
       locations: ['sadad', 'asdad'],
       encodedLocations: '000x9',
-    };
-    const NO_OP_TB_DATA: TrustBoundaryData = {
-      locations: ['sadad', 'asdad'],
-      encodedLocations: '0x0',
     };
 
     function setupTokenNock(email: string | 'default' = 'default'): nock.Scope {
@@ -294,18 +288,6 @@ describe('compute', () => {
           {access_token: MOCK_ACCESS_TOKEN, expires_in: 10000},
           HEADERS,
         );
-    }
-
-    function setupExpiredTokenNock(
-      email: string | 'default' = 'default',
-    ): nock.Scope {
-      const tokenPath =
-        email === 'default'
-          ? `${BASE_PATH}/instance/service-accounts/default/token`
-          : `${BASE_PATH}/instance/service-accounts/${email}/token`;
-      return nock(HOST_ADDRESS)
-        .get(tokenPath)
-        .reply(200, {access_token: MOCK_ACCESS_TOKEN, expires_in: -1}, HEADERS);
     }
 
     function setupTrustBoundaryNock(
@@ -333,34 +315,7 @@ describe('compute', () => {
       nock.cleanAll();
     });
 
-    it('should fetch and return trust boundary data successfully', async () => {
-      const compute = new Compute({serviceAccountEmail: SERVICE_ACCOUNT_EMAIL});
-      const scopes = [
-        setupTokenNock(SERVICE_ACCOUNT_EMAIL),
-        setupTrustBoundaryNock(SERVICE_ACCOUNT_EMAIL),
-        mockExample(),
-      ];
-
-      await compute.request({url});
-
-      assert.deepStrictEqual(compute.trustBoundary, EXPECTED_TB_DATA);
-      scopes.forEach(s => s.done());
-    });
-
-    it('getTrustBoundary should return null when default domain is not googleapis.com', async () => {
-      const compute = new Compute({
-        serviceAccountEmail: SERVICE_ACCOUNT_EMAIL,
-        universe_domain: 'abc.com',
-      });
-      const scopes = [setupTokenNock(SERVICE_ACCOUNT_EMAIL), mockExample()];
-
-      await compute.request({url});
-
-      assert.deepStrictEqual(compute.trustBoundary, null);
-      scopes.forEach(s => s.done());
-    });
-
-    it('fetchTrustBoundary should use the email from metadataServer if no serviceAccountEmail passed', async () => {
+    it('refreshTrustBoundary should use the email from metadataServer if no serviceAccountEmail passed', async () => {
       const compute = new Compute();
       const fakeEmail = 'fake-default-sa@developer.gserviceaccount.com';
       const metadataStub = sandbox.stub(gcpMetadata, 'instance');
@@ -381,7 +336,7 @@ describe('compute', () => {
       scopes.forEach(s => s.done());
     });
 
-    it('fetchTrustBoundary should fail gcpMetadata call', async () => {
+    it('refreshTrustBoundary should throw when gcpMetadata call fails', async () => {
       const compute = new Compute();
       const scopes = [setupTokenNock()];
 
@@ -398,169 +353,6 @@ describe('compute', () => {
         ),
       );
 
-      scopes.forEach(s => s.done());
-    });
-
-    it('refreshTrustBoundary should throw when no valid access token is passed', async () => {
-      const compute = new Compute({
-        serviceAccountEmail: SERVICE_ACCOUNT_EMAIL,
-      });
-      const scopes = [setupExpiredTokenNock(SERVICE_ACCOUNT_EMAIL)];
-
-      await assert.rejects(
-        compute.request({url}),
-        new RegExp(
-          'TrustBoundary: Error calling lookup endpoint without valid access token',
-        ),
-      );
-      scopes.forEach(s => s.done());
-    });
-
-    it('refreshTrustBoundary should return no-op and not call lookup endpoint in case cachedTrustBoundaries is no-op', async () => {
-      const compute = new Compute({
-        serviceAccountEmail: SERVICE_ACCOUNT_EMAIL,
-      });
-      compute.trustBoundary = {encodedLocations: NoOpEncodedLocations};
-      const scopes = [setupTokenNock(SERVICE_ACCOUNT_EMAIL), mockExample()];
-      const tbScope = setupTrustBoundaryNock(SERVICE_ACCOUNT_EMAIL);
-
-      await compute.request({url});
-      assert.deepStrictEqual(
-        compute.trustBoundary.encodedLocations,
-        NoOpEncodedLocations,
-      );
-      scopes.forEach(s => s.done());
-      assert.strictEqual(tbScope.isDone(), false);
-    });
-
-    it('refreshTrustBoundary should return no-op if response from lookup is no-op', async () => {
-      const compute = new Compute({
-        serviceAccountEmail: SERVICE_ACCOUNT_EMAIL,
-      });
-      const scopes = [
-        setupTokenNock(SERVICE_ACCOUNT_EMAIL),
-        setupTrustBoundaryNock(SERVICE_ACCOUNT_EMAIL, NO_OP_TB_DATA),
-        mockExample(),
-      ];
-
-      await compute.request({url});
-      assert.deepStrictEqual(
-        compute?.trustBoundary?.encodedLocations,
-        NoOpEncodedLocations,
-      );
-      scopes.forEach(s => s.done());
-    });
-
-    it('refreshTrustBoundary should return cached TB if call to lookup fails', async () => {
-      const compute = new Compute({
-        serviceAccountEmail: SERVICE_ACCOUNT_EMAIL,
-      });
-      compute.trustBoundary = EXPECTED_TB_DATA;
-
-      const lookupUrl = SERVICE_ACCOUNT_LOOKUP_ENDPOINT.replace(
-        '{service_account_email}',
-        encodeURIComponent(SERVICE_ACCOUNT_EMAIL),
-      );
-
-      const tbErrorScope = nock(new URL(lookupUrl).origin)
-        .get(new URL(lookupUrl).pathname)
-        .matchHeader('authorization', MOCK_AUTH_HEADER)
-        .replyWithError('Something wrong!');
-
-      const scopes = [
-        setupTokenNock(SERVICE_ACCOUNT_EMAIL),
-        tbErrorScope,
-        mockExample(),
-      ];
-
-      await compute.request({url});
-      assert.deepStrictEqual(compute.trustBoundary, EXPECTED_TB_DATA);
-      scopes.forEach(s => s.done());
-    });
-
-    it('refreshTrustBoundary should throw if call to lookup fails and no cached-TB', async () => {
-      const compute = new Compute({
-        serviceAccountEmail: SERVICE_ACCOUNT_EMAIL,
-      });
-
-      const lookupUrl = SERVICE_ACCOUNT_LOOKUP_ENDPOINT.replace(
-        '{service_account_email}',
-        encodeURIComponent(SERVICE_ACCOUNT_EMAIL),
-      );
-
-      const tbErrorScope = nock(new URL(lookupUrl).origin)
-        .get(new URL(lookupUrl).pathname)
-        .matchHeader('authorization', MOCK_AUTH_HEADER)
-        .replyWithError('Something wrong!');
-
-      const scopes = [setupTokenNock(SERVICE_ACCOUNT_EMAIL), tbErrorScope];
-
-      await assert.rejects(
-        compute.request({url}),
-        new RegExp('TrustBoundary: Failure while getting trust boundaries:'),
-      );
-      scopes.forEach(s => s.done());
-    });
-
-    it('refreshTrustBoundary should throw in case of malformed response from lookup', async () => {
-      const compute = new Compute({
-        serviceAccountEmail: SERVICE_ACCOUNT_EMAIL,
-      });
-      const malformedTBData: TrustBoundaryData = {
-        locations: ['sadad', 'asdad'],
-        encodedLocations: '',
-      };
-      const scopes = [
-        setupTokenNock(SERVICE_ACCOUNT_EMAIL),
-        setupTrustBoundaryNock(SERVICE_ACCOUNT_EMAIL, malformedTBData),
-      ];
-
-      await assert.rejects(
-        compute.request({url}),
-        new RegExp('TrustBoundary: Failure while getting trust boundaries:'),
-      );
-      scopes.forEach(s => s.done());
-    });
-
-    it('getRequestHeaders should attach a trust boundary header in case of valid tb', async () => {
-      const compute = new Compute({serviceAccountEmail: SERVICE_ACCOUNT_EMAIL});
-      const scopes = [
-        setupTokenNock(SERVICE_ACCOUNT_EMAIL),
-        setupTrustBoundaryNock(SERVICE_ACCOUNT_EMAIL),
-      ];
-
-      const reqheaders = await compute.getRequestHeaders();
-
-      assert.deepStrictEqual(
-        reqheaders.get('x-allowed-locations'),
-        EXPECTED_TB_DATA.encodedLocations,
-      );
-      scopes.forEach(s => s.done());
-    });
-
-    it('getRequestHeaders should attach an empty string TB header in case of no_op tb', async () => {
-      const compute = new Compute({serviceAccountEmail: SERVICE_ACCOUNT_EMAIL});
-      const scopes = [
-        setupTokenNock(SERVICE_ACCOUNT_EMAIL),
-        setupTrustBoundaryNock(SERVICE_ACCOUNT_EMAIL, NO_OP_TB_DATA),
-      ];
-
-      const reqheaders = await compute.getRequestHeaders();
-
-      assert.deepStrictEqual(reqheaders.get('x-allowed-locations'), '');
-      scopes.forEach(s => s.done());
-    });
-
-    it('getRequestHeaders should not attach TB header in case of non GDU universe', async () => {
-      const compute = new Compute({
-        serviceAccountEmail: SERVICE_ACCOUNT_EMAIL,
-        universe_domain: 'abc.com',
-      });
-      const scopes = [setupTokenNock(SERVICE_ACCOUNT_EMAIL)];
-
-      const reqheaders = await compute.getRequestHeaders();
-
-      assert.deepStrictEqual(reqheaders.get('x-allowed-locations'), null);
       scopes.forEach(s => s.done());
     });
   });
